크아

+++ CRAZYARCADE/ObjectManager.cpp
@@ -11,12 +11,12 @@ ObjectManager::~ObjectManager()
 
 void ObjectManager::AddObject(Object* obj, OBJTYPE type)
 {
-	objects[(UINT)type].emplace_back(obj);
+	objects[type].emplace_back(obj);
 }
 
 void ObjectManager::Update()
 {
-	for (UINT i = 0; i < (UINT)OBJTYPE::END; ++i)
+	for (int i = 0; i < OBJTYPE::END; ++i)
 	{
 		for (auto& Obj : objects[i])
 			Obj->Update();
@@ -25,7 +25,7 @@ void ObjectManager::Update()
 
 void ObjectManager::LateUpdate()
 {
-	for (UINT i = 0; i < (UINT)OBJTYPE::END; ++i)
+	for (int i = 0; i < OBJTYPE::END; ++i)
 	{
 		for (auto& Obj : objects[i])
 			Obj->LateUpdate();
@@ -35,7 +35,7 @@ void ObjectManager::LateUpdate()
 
 void ObjectManager::Render(HDC hdc)
 {
-	for (UINT i = 0; i < (UINT)OBJTYPE::END; ++i)
+	for (int i = 0; i < OBJTYPE::END; ++i)
 	{
 		for (auto& Obj : objects[i])
 		{
@@ -47,7 +47,7 @@ void ObjectManager::Render(HDC hdc)
 
 void ObjectManager::Release()
 {
-	for (UINT i = 0; i < (UINT)OBJTYPE::END; ++i)
+	for (int i = 0; i < OBJTYPE::END; ++i)
 	{
 		for (auto& Obj : objects[i])
 			objects[i].clear();
